fixup: Prefix all new types with GRPC
